refactor: update json2xml scrip, dependencies and readme
