+++ package.json
@@ -1,18 +1,21 @@
 {
+  "type": "module",
   "devDependencies": {
     "@eslint/js": "^9.5.0",
     "ajv": "^8.17.1",
     "ajv-formats": "^3.0.1",
+    "axios": "^1.7.2",
     "eslint": "^9.5.0",
     "eslint-config-prettier": "^9.1.0",
     "globals": "^15.6.0",
     "jest": "^29.7.0",
-    "prettier": "^3.3.1",
-    "axios": "^1.7.2"
+    "jstoxml": "^5.0.2",
+    "prettier": "^3.3.1"
   },
   "scripts": {
     "lint": "eslint .",
     "format": "npx prettier . --write",
-    "test": "jest"
+    "test": "jest",
+    "json2xml": "node ./scripts/json2xml.js"
   }
 }
