--- scripts/json2xml/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "json2xml",
-  "version": "1.0.0",
-  "description": "A script to convert JSON files to XML",
-  "main": "script.js",
-  "type": "module",
-  "scripts": {
-    "convert": "node script.js"
-  },
-  "dependencies": {
-    "jstoxml": "^5.0.2"
-  }
-}
